+++ docs/source/conf.py
@@ -1,7 +1,7 @@
 project = "flask-inputfilter"
 copyright = "2025, Leander Cain Slotosch"
 author = "Leander Cain Slotosch"
-release = "0.2.0"
+release = "0.3.0"
 
 extensions = ["sphinx_rtd_theme"]